requirements.txt
flask>=2.0
openai
pydub
imageio-ffmpeg
python-dotenv
gunicorn
